requirements.txt
requests>=2.25.0
pandas>=1.3.0
supabase>=0.7.1
python-dotenv>=0.19.0
